chore: improve code
