+++ packages/ai-native/src/browser/components/ChatReply.tsx
@@ -22,8 +22,8 @@ import { IIconService } from '@opensumi/ide-theme';
 import { IMarkdownString, MarkdownString } from '@opensumi/monaco-editor-core/esm/vs/base/common/htmlContent';
 
 import { IChatAgentService, IChatContent, IChatResponseProgressFileTreeData } from '../../common';
-import { AIChatService } from '../chat/ai-chat.service';
 import { ChatRequestModel } from '../chat/chat-model';
+import { ChatService } from '../chat/chat.service';
 import { EMsgStreamStatus, MsgStreamManager } from '../model/msg-stream-manager';
 import { IChatAgentViewService } from '../types';
 
@@ -153,7 +153,7 @@ export const ChatReply = (props: IChatReplyProps) => {
   const aiReporter = useInjectable<IAIReporter>(IAIReporter);
   const iconService = useInjectable<IIconService>(IIconService);
   const contextKeyService = useInjectable<IContextKeyService>(IContextKeyService);
-  const aiChatService = useInjectable<AIChatService>(AIChatService);
+  const aiChatService = useInjectable<ChatService>(ChatService);
   const chatAgentService = useInjectable<IChatAgentService>(IChatAgentService);
 
   const isLastReply = msgStreamManager.currentSessionId === relationId;
