+++ packages/ai-native/src/browser/components/StreamMsg.tsx
@@ -2,8 +2,8 @@ import hljs from 'highlight.js';
 import React, { useCallback, useEffect } from 'react';
 
 import { DisposableCollection, IAIReporter, useInjectable } from '@opensumi/ide-core-browser';
+import { localize } from '@opensumi/ide-core-common';
 
-import { AIResponseTips } from '../../common';
 import { EMsgStreamStatus, IMsgStreamChoices, MsgStreamManager } from '../model/msg-stream-manager';
 
 import styles from './components.module.less';
@@ -100,7 +100,7 @@ export const StreamMsgWrapper = (props: IStreamMsgWrapperProps) => {
     () => (
       <div className={styles.ai_chat_code_wrapper}>
         <div className={styles.render_text}>
-          {isError ? <span>{AIResponseTips.ERROR_RESPONSE}</span> : renderContent(content)}
+          {isError ? <span>{localize('aiNative.chat.error.response')}</span> : renderContent(content)}
         </div>
       </div>
     ),
