+++ packages/ai-native/src/common/index.ts
@@ -1,8 +1,9 @@
 import { CancellationToken, Event, FileType, IDisposable, IMarkdownString, Uri } from '@opensumi/ide-core-common';
 
-export const AIInlineChatContentWidget = 'AI-inline-chat-content-widget';
 export const IAINativeService = Symbol('IAINativeService');
-export const AI_CHAT_CONTAINER_VIEW_ID = 'ai_chat';
+
+export const AIInlineChatContentWidget = 'AI_Inline_Chat_Content_Widget';
+export const AI_CHAT_CONTAINER_VIEW_ID = 'AI_Chat';
 
 export interface IChatMessageStructure {
   /**
@@ -83,219 +84,7 @@ export enum AISerivceType {
 }
 
 export const IAIChatService = Symbol('IAIChatService');
-
-export interface PromptOption {
-  language?: string;
-  useCot?: boolean;
-}
-
-export namespace AIResponseTips {
-  export const ERROR_RESPONSE = '当前与我互动的人太多，请稍后再试，感谢您的理解与支持';
-
-  export const STOP_IMMEDIATELY = '我先不想了，有需要可以随时问我';
-
-  export const NOTFOUND_COMMAND = '很抱歉，暂时未找到可立即执行的命令。';
-
-  export const NOTFOUND_COMMAND_TIP = '你可以打开命令面板搜索相关操作或者重新提问。';
-}
-
-// highlight.js/lib/index.js
-export const highLightLanguageSupport = [
-  '1c',
-  'abnf',
-  'accesslog',
-  'actionscript',
-  'ada',
-  'angelscript',
-  'apache',
-  'applescript',
-  'arcade',
-  'arduino',
-  'armasm',
-  'xml',
-  'asciidoc',
-  'aspectj',
-  'autohotkey',
-  'autoit',
-  'avrasm',
-  'awk',
-  'axapta',
-  'bash',
-  'basic',
-  'bnf',
-  'brainfuck',
-  'c-like',
-  'c',
-  'cal',
-  'capnproto',
-  'ceylon',
-  'clean',
-  'clojure',
-  'clojure-repl',
-  'cmake',
-  'coffeescript',
-  'coq',
-  'cos',
-  'cpp',
-  'crmsh',
-  'crystal',
-  'csharp',
-  'csp',
-  'css',
-  'd',
-  'markdown',
-  'dart',
-  'delphi',
-  'diff',
-  'django',
-  'dns',
-  'dockerfile',
-  'dos',
-  'dsconfig',
-  'dts',
-  'dust',
-  'ebnf',
-  'elixir',
-  'elm',
-  'ruby',
-  'erb',
-  'erlang-repl',
-  'erlang',
-  'excel',
-  'fix',
-  'flix',
-  'fortran',
-  'fsharp',
-  'gams',
-  'gauss',
-  'gcode',
-  'gherkin',
-  'glsl',
-  'gml',
-  'go',
-  'golo',
-  'gradle',
-  'groovy',
-  'haml',
-  'handlebars',
-  'haskell',
-  'haxe',
-  'hsp',
-  'htmlbars',
-  'http',
-  'hy',
-  'inform7',
-  'ini',
-  'irpf90',
-  'isbl',
-  'java',
-  'javascript',
-  'jboss-cli',
-  'json',
-  'julia',
-  'julia-repl',
-  'kotlin',
-  'lasso',
-  'latex',
-  'ldif',
-  'leaf',
-  'less',
-  'lisp',
-  'livecodeserver',
-  'livescript',
-  'llvm',
-  'lsl',
-  'lua',
-  'makefile',
-  'mathematica',
-  'matlab',
-  'maxima',
-  'mel',
-  'mercury',
-  'mipsasm',
-  'mizar',
-  'perl',
-  'mojolicious',
-  'monkey',
-  'moonscript',
-  'n1ql',
-  'nginx',
-  'nim',
-  'nix',
-  'node-repl',
-  'nsis',
-  'objectivec',
-  'ocaml',
-  'openscad',
-  'oxygene',
-  'parser3',
-  'pf',
-  'pgsql',
-  'php',
-  'php-template',
-  'plaintext',
-  'pony',
-  'powershell',
-  'processing',
-  'profile',
-  'prolog',
-  'properties',
-  'protobuf',
-  'puppet',
-  'purebasic',
-  'python',
-  'python-repl',
-  'q',
-  'qml',
-  'r',
-  'reasonml',
-  'rib',
-  'roboconf',
-  'routeros',
-  'rsl',
-  'ruleslanguage',
-  'rust',
-  'sas',
-  'scala',
-  'scheme',
-  'scilab',
-  'scss',
-  'shell',
-  'smali',
-  'smalltalk',
-  'sml',
-  'sqf',
-  'sql_more',
-  'sql',
-  'stan',
-  'stata',
-  'step21',
-  'stylus',
-  'subunit',
-  'swift',
-  'taggerscript',
-  'yaml',
-  'tap',
-  'tcl',
-  'thrift',
-  'tp',
-  'twig',
-  'typescript',
-  'vala',
-  'vbnet',
-  'vbscript',
-  'vbscript-html',
-  'verilog',
-  'vhdl',
-  'vim',
-  'x86asm',
-  'xl',
-  'xquery',
-  'zephir',
-];
-
 export const IChatManagerService = Symbol('IChatManagerService');
-
 export const IChatAgentService = Symbol('IChatAgentService');
 
 export interface IChatAgentService {
